[ fix ] api 통합
