+++ components/mypage/index.tsx
@@ -13,7 +13,7 @@ import useModal from '@/hooks/common/useModal';
 import { CakeProfileImg } from '@/public/assets/images';
 import { useEffect, useState } from 'react';
 import useInitEditWishesInfo from '@/hooks/mypage/useInitEditWishesInfo';
-import { deleteUserInfo } from '@/api/mypage/deleteUserInfo';
+import { deleteUserInfo } from '@/api/mypage/mypageAPI';
 
 export default function MyPageContainer() {
   const { isOpen, handleToggle } = useModal();
